requirements.txt
torch>=2.1.0
torchvision>=0.16.0
robosuite>=1.5.0
faiss-cpu>=1.8.0
transformers>=4.35.0
diffusers>=0.25.0
h5py>=3.10.0
numpy>=1.24.0
matplotlib>=3.7.0
PyYAML>=6.0
pyzmq>=25.0.0
opencv-python>=4.8.0
